requirements.txt

docx==0.2.4
faiss-cpu==1.7.4
google-ai-generativelanguage==0.4.0
google-api-core==2.17.0
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.3.2
kaleido==0.2.1
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.30
langchain-google-genai==0.0.9
langchain-openai==0.0.8
langsmith==0.1.8
libclang==16.0.6
llama-index==0.10.30
llama-index-agent-openai==0.2.2
llama-index-cli==0.1.12
llama-index-core==0.10.30
llama-index-embeddings-openai==0.1.8
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.16
llama-index-multi-modal-llms-openai==0.1.5
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.19
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.1
llamaindex-py-client==0.1.18
openai==1.33.0
pandas==2.2.0
plotly==5.3.1